Django/requirements.txt
Django==2.2.7
django-cors-headers==3.2.0
django-elasticsearch-dsl==7.1.0
django-elasticsearch-dsl-drf==0.20.3
django-nine==0.2.2
djangorestframework==3.10.3
elasticsearch==7.1.0
elasticsearch-dsl==7.1.0
python-dateutil==2.8.1
pytz==2019.3
six==1.13.0
sqlparse==0.3.0
urllib3==1.25.7
djangorestframework-jwt==1.11.0